+++ src/forms/__tests__/formVisibility.test.ts
@@ -1,140 +1,6 @@
 import { describe, expect, it } from "vitest";
-import {
-  evaluateRule,
-  resolveFormVisibility,
-  type VisibilityRule,
-} from "../formVisibility";
+import { resolveFormVisibility } from "../formVisibility";
 import { makeStep } from "./testHelpers";
-
-describe("evaluateRule", () => {
-  describe("equals", () => {
-    it.each<[VisibilityRule, Record<string, unknown>, boolean]>([
-      [
-        { field: "hasUsedOtherNameOrAlias", equals: true },
-        { hasUsedOtherNameOrAlias: true },
-        true,
-      ],
-      [
-        { field: "birthplaceCountry", equals: "US" },
-        { birthplaceCountry: "US" },
-        true,
-      ],
-      [
-        { field: "hasUsedOtherNameOrAlias", equals: true },
-        { hasUsedOtherNameOrAlias: false },
-        false,
-      ],
-      [
-        { field: "birthplaceCountry", equals: "US" },
-        { birthplaceCountry: "CA" },
-        false,
-      ],
-      [{ field: "hasUsedOtherNameOrAlias", equals: true }, {}, false],
-    ])("evaluates %s with %o -> %s", (rule, data, expected) => {
-      expect(evaluateRule(rule, data)).toBe(expected);
-    });
-  });
-
-  describe("notEquals", () => {
-    it.each<[VisibilityRule, Record<string, unknown>, boolean]>([
-      [
-        { field: "isCurrentlyUnhoused", notEquals: true },
-        { isCurrentlyUnhoused: false },
-        true,
-      ],
-      [
-        { field: "isCurrentlyUnhoused", notEquals: true },
-        { isCurrentlyUnhoused: undefined },
-        true,
-      ],
-      [
-        { field: "isCurrentlyUnhoused", notEquals: true },
-        { isCurrentlyUnhoused: true },
-        false,
-      ],
-    ])("evaluates %s with %o -> %s", (rule, data, expected) => {
-      expect(evaluateRule(rule, data)).toBe(expected);
-    });
-  });
-
-  describe("includes", () => {
-    it.each<[VisibilityRule, Record<string, unknown>, boolean]>([
-      [
-        { field: "pronouns", includes: "other" },
-        { pronouns: ["they/them", "other"] },
-        true,
-      ],
-      [
-        { field: "pronouns", includes: "other" },
-        { pronouns: ["they/them", "she/her"] },
-        false,
-      ],
-      [{ field: "pronouns", includes: "other" }, {}, false],
-    ])("evaluates %s with %o -> %s", (rule, data, expected) => {
-      expect(evaluateRule(rule, data)).toBe(expected);
-    });
-  });
-
-  describe("and", () => {
-    it.each<[VisibilityRule, Record<string, unknown>, boolean]>([
-      [
-        {
-          and: [
-            { field: "isCurrentlyUnhoused", notEquals: true },
-            { field: "isMailingAddressDifferentFromResidence", equals: true },
-          ],
-        },
-        {
-          isCurrentlyUnhoused: false,
-          isMailingAddressDifferentFromResidence: true,
-        },
-        true,
-      ],
-      [
-        {
-          and: [
-            { field: "isCurrentlyUnhoused", notEquals: true },
-            { field: "isMailingAddressDifferentFromResidence", equals: true },
-          ],
-        },
-        {
-          isCurrentlyUnhoused: true,
-          isMailingAddressDifferentFromResidence: true,
-        },
-        false,
-      ],
-    ])("evaluates %s with %o -> %s", (rule, data, expected) => {
-      expect(evaluateRule(rule, data)).toBe(expected);
-    });
-  });
-
-  describe("or", () => {
-    it.each<[VisibilityRule, Record<string, unknown>, boolean]>([
-      [
-        {
-          or: [
-            { field: "hasUsedOtherNameOrAlias", equals: true },
-            { field: "hasPreviousNameChange", equals: true },
-          ],
-        },
-        { hasUsedOtherNameOrAlias: false, hasPreviousNameChange: true },
-        true,
-      ],
-      [
-        {
-          or: [
-            { field: "hasUsedOtherNameOrAlias", equals: true },
-            { field: "hasPreviousNameChange", equals: true },
-          ],
-        },
-        { hasUsedOtherNameOrAlias: false, hasPreviousNameChange: false },
-        false,
-      ],
-    ])("evaluates %s with %o -> %s", (rule, data, expected) => {
-      expect(evaluateRule(rule, data)).toBe(expected);
-    });
-  });
-});
 
 describe("resolveFormVisibility", () => {
   describe("steps (visibleStepIds)", () => {
@@ -221,7 +87,7 @@ describe("resolveFormVisibility", () => {
       const stepWithVisibility = makeStep("a", [
         "hasUsedOtherNameOrAlias",
         {
-          name: "otherNamesOrAliases",
+          id: "otherNamesOrAliases",
           when: { field: "hasUsedOtherNameOrAlias", equals: true },
         },
       ]);
@@ -258,24 +124,24 @@ describe("resolveFormVisibility", () => {
       ] as const;
       const result = resolveFormVisibility([], {}, pdfs);
       expect(result.pdfsToInclude).toEqual([
-        { pdfId: "cjp27-petition-to-change-name-of-adult", include: true },
-        { pdfId: "ss5-application-for-social-security-card", include: true },
+        "cjp27-petition-to-change-name-of-adult",
+        "ss5-application-for-social-security-card",
       ]);
     });
 
     it.each([
       [
         { shouldApplyForFeeWaiver: true },
-        [{ pdfId: "affidavit-of-indigency", include: true }],
+        ["affidavit-of-indigency"],
       ],
       [
         { shouldApplyForFeeWaiver: false },
-        [{ pdfId: "affidavit-of-indigency", include: false }],
+        [{ id: "affidavit-of-indigency", when: false }],
       ],
     ] as const)("evaluates when rule with formData (formData=%o -> pdfsToInclude)", (formData, expected) => {
       const pdfs = [
         {
-          pdfId: "affidavit-of-indigency" as const,
+          id: "affidavit-of-indigency" as const,
           when: { field: "shouldApplyForFeeWaiver" as const, equals: true },
         },
       ];
@@ -300,7 +166,7 @@ describe("resolveFormVisibility", () => {
       const pdfs = [
         "cjp27-petition-to-change-name-of-adult",
         {
-          pdfId: "affidavit-of-indigency",
+          id: "affidavit-of-indigency",
           when: { field: "shouldApplyForFeeWaiver", equals: true },
         },
       ] as const;
@@ -322,8 +188,8 @@ describe("resolveFormVisibility", () => {
         { stepId: "fee-waiver", fields: ["reasonToWaivePublication"] },
       ]);
       expect(result.pdfsToInclude).toEqual([
-        { pdfId: "cjp27-petition-to-change-name-of-adult", include: true },
-        { pdfId: "affidavit-of-indigency", include: true },
+        "cjp27-petition-to-change-name-of-adult",
+        "affidavit-of-indigency",
       ]);
     });
   });
