+++ src/pages/forms/social-security/_steps/PreviousSocialSecurityCardStep.tsx
@@ -5,7 +5,7 @@ import {
 } from "@/components/react/forms/FormStep";
 import { NameField } from "@/components/react/forms/NameField";
 import { YesNoField } from "@/components/react/forms/YesNoField";
-import type { VisibilityRule } from "@/forms/formVisibility";
+import type { VisibilityRule } from "@/forms/visibilityRules";
 import type { Step } from "@/forms/types";
 
 const whenHasPreviousCard: VisibilityRule = {
@@ -21,7 +21,7 @@ export const previousSocialSecurityCardStep: Step = {
   fields: [
     "hasPreviousSocialSecurityCard",
     {
-      names: [
+      ids: [
         "previousSocialSecurityCardFirstName",
         "previousSocialSecurityCardMiddleName",
         "previousSocialSecurityCardLastName",
