+++ src/pdfs/utils/loadPdfs.ts
@@ -4,25 +4,25 @@ import { getPdfDefinition } from "@/pdfs";
 /**
  * Load multiple PDFs.
  *
- * @param pdfs - Array of PDF configurations
- * @param pdfs[].pdfId - The ID of the PDF to load
- * @param pdfs[].include - Whether to include this PDF (defaults to `true`)
+ * @param pdfs - Array of PDF configurations. Id alone = include; { id, when: false } = exclude.
  * @returns Promise that resolves to an array of PDF definitions
  *
  * @example
  * ```typescript
  * const pdfs = await loadPdfs([
- *   { pdfId: "always-included-pdf" },
- *   { pdfId: "conditional-pdf", include: someCondition },
+ *   "always-included-pdf",
+ *   { id: "conditional-pdf", when: false },
  * ]);
  * ```
  */
 export async function loadPdfs(
-  pdfs: Array<{ pdfId: PDFId; include?: boolean }>,
+  pdfs: readonly (PDFId | { id: PDFId; when?: boolean })[],
 ) {
-  const pdfIds = pdfs
-    .filter(({ include = true }) => include)
-    .map(({ pdfId }) => pdfId);
+  const ids = pdfs
+    .filter((entry) =>
+      typeof entry === "string" ? true : entry.when !== false,
+    )
+    .map((entry) => (typeof entry === "string" ? entry : entry.id));
 
-  return await Promise.all(pdfIds.map((pdfId) => getPdfDefinition(pdfId)));
+  return await Promise.all(ids.map((id) => getPdfDefinition(id)));
 }
